Migrates part of config node to use Socket
