+++ test/test_notecard_node.js
@@ -6,6 +6,7 @@ const helper = require("node-red-node-test-helper");
 const ncNode = require('../notecard/notecard-node.js');
 const transactor = require('./inmem_transactor.js');
 
+
 const NotecardMessageTerminator = '\r\n';
 
 class BusMockTransactor extends transactor.InMemTransactor{
@@ -38,7 +39,7 @@ describe('Notecard Config Node', function() {
     afterEach(function() {
         helper.unload();
     });
-    const flow = [{ id: "n1", type: "notecard-config", name: "Notecard Config" }];
+    const flow = [{ id: "n1", type: "notecard-config", name: "Notecard Config", other:"prop" }];
 
     it('should be loaded with correct default values', function(done) {
         
@@ -57,6 +58,8 @@ describe('Notecard Config Node', function() {
         
     });
 
+    
+
     const loadFlow = (node, flow) => {
         const p = new Promise((resolve, reject) => {
             try {
@@ -70,6 +73,49 @@ describe('Notecard Config Node', function() {
 
         return p;
     }
+
+    const loadAndGetNode = async (f) => {
+        if(f === undefined)
+            f = {};
+        f.id = "n1";
+        f.type = "notecard-config";
+        f.name = "Notecard Config";
+        await loadFlow(ncNode, [f]);
+        return helper.getNode(f.id);
+    }
+
+    it('should have config values', async () => {
+        const name = 'myConfigProp'
+        const value = 'myConfigValue'
+        let c = {};
+        c[name] = value;
+        const n1 = await loadAndGetNode(c)
+        
+        const config = n1.Config
+
+        config.should.have.property(name,value)
+
+    });
+
+    describe('Notecard Socket Config', () => {
+        it('should configure UART socket with Serial Config options', async () => {
+            const config = {socket:'uart', port:'abc', baudrate: 115200};
+            const n1 = await loadAndGetNode(config);
+
+            n1.Notecard.Socket.should.have.property('Port', config.port);
+            n1.Notecard.Socket.should.have.property('BaudRate', config.baudrate);
+        });
+
+        it('should apply arbitrary socket object if socket property is an object', async () => {
+            socket = {name:"value"};
+            const config = {socket:socket}
+
+            const n1 = await loadAndGetNode(config);
+
+            n1.Notecard.Socket.should.deepEqual(socket);
+
+        });
+    });
 
     it('should apply node bus number field value to transactor if populated', async () => {
         var transactor = new BusMockTransactor();
